pyproject.toml
[tool.poetry]
name = "microblog-service"
version = "0.1.0"
description = "\"Python Advanced\" course diploma"
authors = ["https://github.com/Griy-dot-py"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.29.0"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
uvicorn = "^0.30.1"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.1"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
psycopg2-binary = "^2.9.9"
ruff = "^0.4.10"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
